DBG: fixed issue #112 (binary search in memory map over all pages)

+++ x64_dbg_dbg/filehelper.cpp
@@ -25,5 +25,5 @@ bool FileHelper::WriteAllText(const String & fileName, const String & content)
     if(hFile == INVALID_HANDLE_VALUE)
         return false;
     DWORD written = 0;
-    return !!WriteFile(hFile, content.c_str(), content.length(), &written, nullptr);
+    return !!WriteFile(hFile, content.c_str(), DWORD(content.length()), &written, nullptr);
 }
